Change create conversation layout
To accomplish that, each page needs to build its own layout from scratch. Implementing different shared layouts is an option to increase reusability.
