+++ client/src/pages/Login.js
@@ -10,6 +10,7 @@ import { Alert, Paper } from '@mui/material'
 import axios from 'axios'
 import { baseUrl } from '../axiosConfig'
 import ProtectedRoute from '../components/ProtectedRoute'
+import Navbar from '../components/Navbar'
 
 function Login() {
 	const navigate = useNavigate()
@@ -57,51 +58,60 @@ function Login() {
 	}
 	console.log(message)
 	return (
-		<Box 
-			sx={{
-				display: "flex",
-				flexDirection: "column",
-				justifyContent: "space-between",
-				alignItems: "center",
-			}}
-		>   
-			<form onSubmit={ handleSubmit }>
-				<Box 
-					sx={{
-						display: "flex",
-						flexDirection: "column",
-						justifyContent: "center",
-						alignItems: "center",
-						gap: "2vh"
-					}}
-				>
-					<Typography fontWeight="100" sx={{mb: 5}} variant="h3">Login</Typography>
-					{ message && <Alert sx={{m: 2}} variant="filled" severity={message.type}> { message.text } </Alert> }
-					<TextField 
-						onChange={handleChange} 
-						value={formData.username} 
-						autoComplete="off" 
-						name="username"
-						label="Username"
-						variant="outlined"
-						InputProps={{ sx: { borderRadius: '25px' }}}
-					/>
-					{/* <Typography variant="body2">Password:</Typography> */}
-					<TextField
-						onChange={handleChange} 
-						value={formData.password} 
-						autoComplete="off" 
-						name="password"
-						label="Password"
-						type="password"
-						variant="outlined"
-						InputProps={{ sx: { borderRadius: '25px' }}}
-					/>
-					<Button sx={{ width: '100%', borderRadius: '25px', mt: 2, display: "block" }} type="submit" variant="contained">Enter chat</Button>
-					<Button sx={{ width: '100%', borderRadius: '25px'}} variant="outlined" onClick={() => {navigate('/signup')}}>Sign up</Button>
-				</Box>
-			</form>
-		</Box>
+		<Paper square sx={{ 
+			display: "flex", 
+			flexDirection: "column", 
+			width: "100vw", 
+			height: "100vh", 
+			padding: '2vh 5vw' 
+		}}>
+			<Navbar />
+			<Box 
+				sx={{
+					display: "flex",
+					flexDirection: "column",
+					justifyContent: "space-between",
+					alignItems: "center",
+				}}
+			>   
+				<form onSubmit={ handleSubmit }>
+					<Box 
+						sx={{
+							display: "flex",
+							flexDirection: "column",
+							justifyContent: "center",
+							alignItems: "center",
+							gap: "2vh"
+						}}
+					>
+						<Typography fontWeight="100" sx={{mb: 5}} variant="h3">Login</Typography>
+						{ message && <Alert sx={{m: 2}} variant="filled" severity={message.type}> { message.text } </Alert> }
+						<TextField 
+							onChange={handleChange} 
+							value={formData.username} 
+							autoComplete="off" 
+							name="username"
+							label="Username"
+							variant="outlined"
+							InputProps={{ sx: { borderRadius: '25px' }}}
+						/>
+						{/* <Typography variant="body2">Password:</Typography> */}
+						<TextField
+							onChange={handleChange} 
+							value={formData.password} 
+							autoComplete="off" 
+							name="password"
+							label="Password"
+							type="password"
+							variant="outlined"
+							InputProps={{ sx: { borderRadius: '25px' }}}
+						/>
+						<Button sx={{ width: '100%', borderRadius: '25px', mt: 2, display: "block" }} type="submit" variant="contained">Enter chat</Button>
+						<Button sx={{ width: '100%', borderRadius: '25px'}} variant="outlined" onClick={() => {navigate('/signup')}}>Sign up</Button>
+					</Box>
+				</form>
+			</Box>
+		</Paper>
 	)
 }
